Phase 2: Unified board screen — fold practice mode into GamePlayView, delete dead code
Consolidate PracticeSessionViewModel/PracticeOpeningView into the existing
GamePlayViewModel + GamePlayView architecture. Move shared session types
(BookStatus, SessionMode, SessionStats, ExplainContext) to SessionTypes.swift.
Delete 6 dead files (~3,250 lines): SessionViewModel, SessionView,
PracticeSessionViewModel, PracticeOpeningView, SessionFeedManager,
SessionProgressTracker.
